Have paparazzi listen for updates and behave as read-only client.
Also creates a new tmz service which manages the paparazzi. This involves posting messages to a service bus queue to indicate snapshots are required for a given document.

The paparazzi service will listen for messages on this queue and when they exist will start up and behave as a read-only client.