Roll back storing offset in message and just use sequence number
We introduced the offset previously to be able to determine messages that need an ACK to move the min sequence number from those that don't. But this extra offset adds complexity for the collaborative types. Especially the merge tree which snapshots on the min sequence number. It would need to be augmented to also track the offset for this number.

Instead it's simpler to update the delta manager to not ack certain messages - like those that change the min sequence number. This also gives us one number to track which identifiers deltas.