Fix for deli ticketing assert. (#237)
Fixes an assert bug coming from an incorrect invariant assumption. The MSN computed from clients is not guaranteed to always increase. Since a new client may come in greater than the published MSN (which lags the one computed directly off clients). This change updates the code to handle this case.