Uninstall LFS (#797)
GitHub puts a bandwidth cap on LFS which we share with all of Microsoft. If the cap is hit LFS fails. The same cap doesn't apply to the repo. Given we are only storing a small number of images and these images don't rev often and are included in the base checkout anyway for simplicity we will just disable LFS.