Reenable localTestServer test (#1628)
* Reenable localTestServer test

The timing issue should be fixed by PR #1528 where we me make sure the
pending work in server is drained.

Refactor TestResolver the mapEndToEnd test to be used in other test as
well.

Fix test result copying in build:copyresult after the package directory structure changed so the build pipeline have the reports to analyze

* Remove unused package dependency

local-test-server should not depend on routerlicious