Single page app with routing chaincode sample (#1688)
Pulls in the Angular 2 tour of heroes sample app which creates a single page app with multiple routes. In this app you have a management dashboard for a set of superheroes. Each has details that can be drilled in to. But this dashboard defines a url hierarchy - i.e.

/dashboard
/detail/:heroId
/heroes

But the rendering of this hierarchy is fully controlled by the single page app.

It is interesting as a chaincode because through our request handling we can insert the hero details directly into another document and use the angular route handler to display the details page.