Update tour of heroes to be able to embed direct link to hero (#1690)
The tour of heroes container now defines a custom request URL handler which enables a hero page to be embedded as a component in another document. This involves more manual management of Angular's default router behavior which assumes full page access and a desire to directly tweak the page URL itself. We still allow this when the tour of heroes chaincode owns the full document. But disable it when embedding a direct link to a hero.