Add btree.ts to map package to start creating virtual object collection. (#1732)
We will be using btrees to support virtual object collections. The btree will be the in-memory representation of an indexed object collection. It may represent portions of the collection that are not paged into the client. We will use the btree to create incremental snapshots, to reduce per-snapshot size and to support loading of an index during snapshot load.

This set of commits adds the initial btree.ts implementation,