Consensus Ordered Collection (Resolves #1422) (#1777)
An consensus ordered collection is a distributed data structure, which
holds a collection of JSON-able or distributed objects, and have a
deterministic add/remove order.

The order the server receive the add/remove operation determines the
order those operation are applied to the collection. Different clients
issuing `add` or `remove` operations at the same time will be sequenced.
The order dictates which `add` is done first, thus determining the order
in which it appears in the collection.  It also determines which client
will get the first removed item, etc. All operations are asynchronous.
A function `wait` is provided to wait for and remove an entry in the
collection.

All non-distributed object added to the collection will be cloned (via JSON).
They will not be references to the original input object.  Thus changed to
the input object will not reflect the object in the collection.

Currently, ConsensusQueue and ConsensusStack are two concrete
ordered implemented. Additional collection order can be added by
implementing the IOrderCollection for add/remove and snapshotting,
and pass it into ConsensusOrderedCollection as the backing data store.