Consolidate chaincode shared-text and controller shared-text (#2039)
For prototyping we had forked shared text that exists inside of routerlicious into the chaincode folder. Things have stabilized enough that we can now bring them back together.

This change also updates shared-text to work as a local chaincode package - just inside of the routerlicious folder. This will allow us to move gateway to be a loader only host.