Initial documentation setup (#1905)
This PR lays the groundwork for reference and narrative documentation as a first step to address #58.

Narrative documentation is in /docs, and reference docs for the API can be generated by running npm run build:docs in the routerlicious folder. Full instructions are in docs/building-documentation.md.

This PR does not automate the documentation build or include it in any of the standard build tasks (e.g. npm run build does not also build documentation). Those changes will be made in a future PR.

Only the @prague/consensus-ordered-collection package is currently documented; others will be added in future PRs.