Remove legacy IChaincode interfaces (#2083)
Removes the legacy definitions inside of @prague/runtime-defintions. This primarily was the IChaincode and IRuntime interfaces. These interfaces existed to make use of the legacy chaincode packages as the component runtime was bootstrapped. Now that these legacy packages have been converted to the updated API there is no longer a need to have these legacy interfaces in the core runtime.