Moving packages out of routerlicious/move routerlicious to server (#2133)
The packages can be built locally via npm scripts at the root. server/routerlicious will use docker to build the packages as well (on top of composing the server)

Also moved shared/test-utils and shared/tools-core to server as they are only used by the server.

packages/server will eventually move out in the future when we are ready to independently rev the versions