Summary op API interfaces and implementation (#2325)
The API should still be considered experimental and snapshots should continue to be used.

Adds in a set of interfaces to define and generate summaries and their associated summary op. And then updates scribe to be able to process them.

And some context on summary ops...

The main difference between summaries and snapshots is that the summary of the document referred to by the summary op is required to be stored by a storage service whereas snapshots were considered optional. But the summary and the snapshot will be the exact same data. We like to think of the summary op addition as creating a base distributed data structure that is specified at the protocol level around storing and modifying immutable graphs of data.

In the current system clients generate ops that the service orders and stores. As a performance optimization we periodically create snapshots of the op log at a particular sequence number with the invariant being loading of a snapshot at a particular sequence number is equivalent to having played all ops up to that sequence number. Snapshots are created by first electing a client to make them - either a browser based one or one running in the cloud - followed by that client writing the snapshot to a content addressable file system (git, or antler for SPO) as an immutable tree.

These snapshots are considered optional. So a service is allowed to drop them if needed. That being said we’ve always treated them as pretty fundamental.

This model has served us well for quite awhile but had a couple difficulties. Given the optional nature of the snapshot - and the invariant to keep it in sync with processing the op log - storage partners wanted to be fully in charge of creating them. This meant they had to run dynamic code inside their service which added a level of complexity to the implementation. Right to forget type features also complicate things a bit since there is a desire to trim parts of the history. And then although we called snapshots optional, given the perf benefits, they become pretty fundamental so we end up having two primary storage points - the op log and the snapshots.

We were working through some of the above complexities with SPO and they floated the summary op idea which we really liked.

In the summary op world we introduce a new op - the summary op - which specifies the state of the document at a particular sequence number. Unlike the snapshot which is required to be a consolidated log up to that sequence number - the summary can be whatever the app chooses. 99% of the time this will just end up being what the snapshot was. But in some cases the app may choose to blow away history. Or reset the state of the document. The state of the summary op becomes the truth of the document once it is processed. The op itself will specify modifications to an immutable graph stored in a content addressable file system.

Ops of course still exist in the system. But the storage layer is only required to store the ops that have not been processed as part of a new summary (i.e. any op whose sequence number is > the sequence number of the last summary). Once a new summary is processed the storage service can delete any ops they were storing. If an app needs this data it's up to it to persist the ops within its summary. So in this world the summaries become the primary storage object and the source of truth.

To someone using our runtime this will all end up being pretty transparent. They'll still load off the latest summary, catch up on any pending ops not yet part of a summary, and then begin submitting ops. If they're elected to make the summaries they'll now submit an op with the result of it rather than directly writing to storage. You can still query the full op stream but may need to walk previous summaries to find all the ops. When a new summary arrives, given eventual consistency, clients will almost always ignore it since they know they already have the correct state (although they could choose to verify and re-load if out of sync). But the added flexibility of the summary op simplifies the system a lot (especially for the service) and opens up some cool new features later on.

* handlers for various system messages

* Make snapshot private in container

* Add protocol semver to socket connection messages

Provides simple support to do basic protocol negotiation between a client and server. The client sends an array of protocol semvers it supports.  The server then selects one based upon its own list of supported versions. The server also returns all possible versions in the case the client would like to renegotiate.

* Update scribe to listen for and react to a summary op

* Use document lambda for scribe

* Have scribe take in mongo and storage endpoints

* Summary ops write to storage

* Update scribe to write to mongo

* Track log offset value inside of scribe

* Initial spec of summary op interfaces

* Begin adding in summary op generation APIs

Also updates the pause and resume methods on DeltaManager to return promises. We may be processing a message when a pause method is invoked. The promise indicates when the pause actually has taken effect.

* Update r11s driver to generate summary package

* Create snapshot to summary converter

* Update scribe lambda to update summary op refs