Redesign getDeltasCore to unify error paths, maing it async and using awaits (and getting rid of deferred object). (#2423)
Unifying error paths, making GetDeltasCore async and using awaits (and getting rid of deferred object).

We hit issues in replayP continuation when we hit 404 on the server and SPO driver gives us deltas = undefined and that code paths has different behavior from errors returned as exceptions.
Changing this code to handle all errors the same - retry (and log appropriately in telemetry)
We also do not want to retry if we fail to deltaStorage itself - it's a game over scenario.