Telemetry and socket.io bug fixes (#2464)
* prepareErrorObject fix

A null error object will mark its typeof as object leading to a later exception

* Revert change to document delta connection

A function as a last parm to a socket.io emit call is a RPC endpoint. We follow the common node/JS convention of the first param being an error value. If it is truthy there is an error otherwise the call was successful.