Scribe service durability plus protocol serialization (#2516)
Updates the Scribe lambda to properly checkpoint against Kafka and maintain durability. It also makes it responsible for writing the protocol state to the summary.

The protocol state is defined as the core state shared between the loader and the service. This is the quorum information combined with the sequence number and min sequence number. It provides a snapshot of the ordering service at the sequence number of the summary. By having the service write this we guarantee its correctness and it serves as a point of recovery. The app snapshot is then combined with this information.

To track this system protocol messages (join/leave/propose/reject) are handled by a common class. Scribe and the loader both make use of this. Scribe is responsible for computing and then writing a serialized form of this.

The scribe lambda has been updated to track this information and then write the information necessary to track this as part of its kafka checkpointing process. As always it commits this information first - with the kafka log offset - prior to notifying Kafka that it can checkpoint a log position.

* initial move of common loader code to shared protocol class for use within scribe lambda

* Start of blob management component

* Split shared protocol handling off from base loader

* Have scribe make use of protocol op handler

* Add in explicit summary op

* Append protocol summary to app summary on server

* Track protocol head in git summaries

* Broadcast summarynack message

* Store pending ops