Generate and load summaries (#2566)
Updates the runtime to create and make use of summaries. To ease the transition disables this by default and continues to use the existing snapshot method. But users can opt into this by a set of options on the runtime load.

Once opted in the runtime will elect the lowest client (by join sequence number) as the summarizer. In order to free up the UI thread, as well as known issues with snapshotting data structures with pending ops, the client opens a new, readonly version of the document, QI's against this for the ISummarizer interface, and then asks this to generate summaries. These then go through the service and scribe takes care of appending protocol information to the app snapshot and then broadcasting the existence of the summary.

On load if a .protocol folder exists we load off the summary. Otherwise use the existing snapshot behavior.

* Move partial ops to runtime

* Call context submit function

* Track earliest client and mark them as the summarizer

* Add in summarizer to runtime

* Loader options to not use cache or reconnect on load

* Update runtime to correctly generate summaries

* Read and process summaries

* summary manager bug fix

* Add flag to container runtime to turn on summary generation