Enable summaries by default on flow and flow-view (#2573)
Also lock prague-dev-server to ~3.5.0 vs ^3.5.0. To keep a version somewhat in sync with the webpack-dev-server base we match the major. The policy is our minor matches the Fluid minor. So we currently are on 0.5. The ~ will keep us in the 3.5.* range vs >=3.5.* < 4.0.0 with the ^. With the ~ we will match the 0.5 dev server.