Componentize factory methods (#2661)
* Componentize factory methods

Updates the system such that the result of a dynamic load is an IComponent. Users can then QI for the given factory method (IRuntimeFactory or IComponentFactory). Or as an IComponent this opens up the dynamic element to expose new capabilities (metadata list, preview images, etc...).

To simplify this disccovery process introduces the IFluidModule interface which defines an export named fluidExport. A fluidExport is an IComponent.