Offline transitions (#2623)
* Add 'online' property to network error events. This is not very useful today (as support for online notifications is not there across browsers today) .
Log errors only once for web socket connection, as we get thousands of errors in offline mode
Add tracking of latency for message round-trip (per Gary request).

* Propagating connected state to quorum and properly handling leader election & rejection of local proposals on disconnect.
Before this change we
- would make leadership proposal in Connecting state, causing it to be dropped and not elect a leader
- any proposals that made it to DeltaManager (but not necessarily to server) would hang in the air on disconnect.

* codeLoader.ts calls registerTasks() before we are connected, causing runtime error (because ContainerRuntime.leader() returns true and we happily call into runTaskAnalyzer())

* Addressing regression in logging code - we are no longer collecting stacks or exception messages.
The issue is - these properties are non-enumerable, and thus did not made it to error2 object.

Addressing assert in ContainerRuntime that fire because of NoOp messages sent from DeltaManager.updateSequenceNumber when in connecting state, responding to Propose message.

Leader rejection & attachComponent() component not to submit messages in disconnected state.