Generate snapshots via Replay tool & Load into FluidDebugger, plus bugfixes (#2741)
The goal of this work is to have a tool to visually inspect snapshot round-trip process in actual apps.
I.e. ability to generate snapshot at particular seq# locally, and load it into the app.
As part of working on it, I hit a bunch of problems in replay tool (which I use to generate such snapshots), so majority of fixes are touching it, but there are also changes in shipping code.

Replay Tool:

Added ability to write out snapshot.json - self-contained snapshot file that other web-based tools (like FluidDebugger) can consume. Format: IFileSnapshot
We deadlocked on loading container (load() function), waiting for loader callback to be called. That resulted in 5 ops being processed, and Node process exiting.
Even if we passed that, replayTool.to = -1 as default was resulting in us not making much progress.
Windows paths (containing ":") were not accepted
For some reason (not investigated further), async version of mkdir was not returning for me sometimes (handing process). sync version works fine.
Get rid of start() logic - we do not need to process any ops when starting from seq#0. Same for !container.existing case - it's not needed.
When generating multiple snapshots, we need to server document tree generated for previous snapshot, as ComponentRuntime.snapshot() needs it to generate next snapshot
Smallish improvements to help screen (print it if no args specified, better descriptions).
Changed directory structure. It used to be ms/scriptor/. Now it's /@ms_scriptor.json

FluidDebugger:

Added button to load external snapshot
For now it's rather ugly experience (both UI-wise and functionality-wise) as I do not check that snapshot is for that particular file / app, so ops playback is disabled after that.
This is mostly to validate that snapshots we generate at specific OP can be loaded and process of generating / loading snapshots does not corrupt document.

Shipping code (not tools):

Added logging for any error that happens at Container.load() process, as those errors go to /dev/null (SharedText), making it hard to debug issues
ContainerRuntime would regenerate component tree if for some reason commits are not to be found in latest tree.

PragueDump:

Ports for local server are wrong.

Things that can be better:

Partial snapshots (that replay tool writes out, not new self-contained snapshot.json) do not write out component trees if components did not change from previous snapshot. That might be misleading to users.
summary is written out, but I do not expand commits (i.e. it's not self-contained).
Code (both in FluidDebugger and in File storage) needs refactoring, but I left it for later (to be its own PR).