Generate package and version name in script to be accessible by code (#2806)
Move where we patch the version before the actual build.
Move the script in the pipeline into the repo.
Add gen_version.js to build-common to generate and update the
packageVersion.ts in the source.

Add packages that using `debug` package to output the package and
version name so we can tell what version is running.