Fix Odsp driver without prefetch snapshot and add pragueDump support for shared point hosted URL (#2834)
When Odsp driver doesn't have a prefetch snapshot, it assumes that there are no snapshot at all and return nothing.   But there are usage of the server that doesn't want to do prefetching.

Change it so that if a promise isn't passed in, use a `undefined` latestSha to indicate that we didn't do prefect.  If the promise resolves to `undefined`, we will use `null` latestSha to indicate that there is no
snapshot and skip querying the server.

Also, add support to sharepoint hosted URLs on personal drives.