Delay load components and distributed data structures (#2819)
Creates a context which proxies pending operations to the object. Buffers these operations when not loaded and processes them once loaded.

Updates the component runtime load method to no longer return the runtime but instead provides it to a callback. This callback is invoked once the runtime is fully loaded and caught up to the sequence number.

Developers should register request handlers inside this callback.