Addressing issue with delay loading SharedStrings (#2929)
When we delay load component that has SharedString, MSN passed to it is current MSN, not MSN of snapshot when SharedString was snapshotted.
As result, as we apply tardis & pending message, they come from the past, i.e. their ref seq# < msn. This will result in incorrect application of ops, as zamboni process can nuke proper segments required to apply such ops.