Fix #2596: node should not have negative length (#2997)
When we compute the partialLength for a node from it's child, we try to
merge the overlapRemoveClient.  But we sometime take a reference of the
data from the children (both the RedBlackTree and it's properties), and
modify it, corrupting the information for the children.

The fix is to clone the information for the current node instead of
reusing the object from the children.

With this fix, we can run conflictFarm with > 100k ops