Fixing a bunch of snapshot bugs (cherry pick from master) (#3012)
* Cherry pick commit: 60d078653466cf7af6dd7bcbdf8a10f7d0352710 [abad8e2]
Fixing a bunch of snapshot bugs (#2983)

A bunch of fixes in response to ReplayTool finding bugs in snapshots:
- getValidOpRange() was not using ref seq# & clientID, thus incorrectly mapped ranges.
- Similar, transformOp was not using ref seq# & clientID for insert ops.
- Similar for tardisPosition
- AnotateRage was using incorrect seq for zamboni process,

Got rid of MSN notifications - sequence does not need it, as it updates properly on every op & on snapshot.
Added strict validations around seq numbers, which helped find some of the bugs.
Make zamboni process 4 segments (instead of current 2) to always keep up, and not bail out on first segment with no parent
Proper patching of ref & min seq numbers for messages in tardis
Fixing all the tests who used undefined and out of order seq numbers

WARNING:
not all tardis related issues are addressed :(
In fact, the more I poke at this code, the more I think it's not fixable.
These changes give substantial improvement in reliability, but there are cases where we produce garbage.
Sort of good news - in most of those case we will trip over message.contents assert and fail to generate broken snapshot