Batched message scheduling (#3204)
With prepare gone it's now possible to batch process messages.

When the flush mode is set to manual - indicating that the user wants manual control over batch ordereding messages - tags the first message with a { batch: true} piece of metadata. The runtime then interprets all future messages from that particular client as being processed in the same turn. When the turn ends, or when flush() is called, the runtime listens to a message from the DeltaManager indicating that the batch is about to be flushed (i.e. added to the send queue to the ordering service). It then updates the last message to mark the end of the batch via a { batch: false } flag.

On inbound, if the batch flag is seen, the runtime will pause the message queue until the end of the batch is encountered. Once this occurs it will unpause the message queue. Because messages are processed in a single turn you're guaranteed to then process these messages in one group.

To help apps track batches introduces a batchBegin/batchEnd event on the container runtime. These mark the beginning and ending of a set of batched operations. For simplicity treats automatic mode individual messages as individual batches as well. The messages then signal the beginning and end of a set of messages that are intended to be seen only at the beginning/end of them being processed.

And also updates the dev container with Zsh + some useful plugins.

* Add zsh and oh-my-zsh to dev container

* batched message scheduling

* Delta manager proxy to localize delta manager to container instance

* await pause promise in serializers

* Always emit batch begin/end messages

Don't expose public access to the summary. The current summarize code makes assumptions about the state at which it was called (paused ops, at a batch boundary) that are guaranteed by higher level code.