Fix two TS definition issues in server packages
localOrderer uses a relative path reference to a file in a separate package rather than to the package itself. This causes problems when directly installing it.

Update socket-storage-shared to include typing definitions as a dependency rather than a devDependency. It's a required reference for its own d.ts definitions so needs to be installed with the package.