Add component interface discovery (#63) (#179)
(Moved from Prague pull request https://github.com/microsoft/Prague/pull/3344)

This change adds interfaces to support component interface discovery. The idea is that a component could say it only cares about other components that are `IComponentHTMLRender`, for example, and it will be notified of components that implement that type on its creation and when new components are added to the ecosystem.

Components who want their functionality to be discoverable would implement `IComponentDiscoverableInterfaces`, and list the interfaces they implement. Components that want to be notified of other components that implement certain interfaces would implement `IComponentDiscoverInterfaces`. Ideally we would be able to force components that implement a certain interface to list that interface string in `discoverableInterfaces`, but it doesn't seem that TypeScript supports creating an interface with a property that can be unioned. So while we could enforce a specific component interface to have a `type` property with a `string` of the interface name, like we did in `IData` in the previous iteration, it doesn't seem that we can merge all the `type` properties into a list for a component that implements multiple interfaces. 

To fit producing/consuming into this paradigm, a component would implement an interface `IComponentProduce[DataType]`, i.e. `IComponentProduceString`, and that interface could be implemented as:

```
export interface IComponentProduceString {
    getString(): string;
}
```

This afford us more type safety for the functionality on a component that produces a specific type, instead of passing in `args: any` as before. Some `IComponentProduceXXXX` interfaces may be more complex, and could contain functionality for data binding. For example, `IComponentProduceString` could have a function `onStringChanged(callback: (newString: string) => void): void;` where the callback would be invoked whenever the string is changed.

This change also introduces the interface for an `IComponentTracker` component. We expect a component tracker to be notified of all newly created components, and then the tracker itself will notify `IComponentDiscoverInterfaces` components of any `IComponentDiscoverableInterfaces` components in the ecosystem with functionality that they care about.