Added retry for secondary AFD url for delta stream socket. (#190)
This change allows socket connection to retry on a second url returned by join session, if applicable. The first url is supposed to be a non-AFD url to Push, and the second url is an AFD one. In the event the user can't connect to the non-AFD endpoint (due to firewalls, etc.), it will retry on the second url.