Clients are missing ops on connection, causing long connection delays on SPO (#428)
There is a timeframe when we are already connected to websocket and lower layers unsubscribed early op handlers, but due to async / continuations, DeltaManager did not have a chance to subscribe its own op handler. As result, we are missing ops, and because PUSH flushes them to storage very infrequently, client is sitting there and asking for missing ops for very long time (30+ seconds).

The fix is to keep early handlers until someone is asking for early ops.
This also requires layers in between not to ask for early ops.

Need to follow up with OuterDocumentService.createDocumentDeltaConnection which has same problem.
WS socket implementation has same issue (there is no early op handler at all there), but given code is not currently used, I've added assert and comment making sure we address that problem if we every switch to raw sockets.