Regression in Container close flow (#468)
As part of closing document, we close connection. That results in disconnection, and thus reconnect flow. Due to a bug, we treat this case as critical error, and raise to an app.
In future, we should reconsider if/when DeltaManager.reconnectOnError() should supply error object. For example, when we add ability to pause connection (not reconnect), this workflow should not raise critical error.