Component Registry Improvements (#435)
The major change here is that now ComponentRegistries support containing other ComponentRegistries, rather than just factories. This removes the wrapping anti-pattern where a registry would call a sub registry's get directly which would make the sub registry in-accessible. Since the sub registry was in-accessible it couldn't have other IComponent* extensions. Getting the sub registries is important for a number of scenarios. For instance the registry may provide meta-data about the packages it contains which can be useful for component pickers. A specific package registry could also provide features like copy/paste and upgrade.

This also results in fewer registries being created, as most cases just wanted to propagate entries for the registry, which is now also supported in the first class way, and is backwards compatible with maps as registries, which is by far the largest use case.

The patterns for creating true sub registries is now to create them explicitly, and with a name; just like any other entry. The fits nicely with the component sub-registry pattern. For example, a cdn registry could be named "cdn", and have it's entry added to the container regsitry, ["cdn", Promise.Resolve(new CdnRegistry())]. Components created from this regisitry would then have a package path like ["cdn", "@some-scope/some-pkg@some-version", "some-component"]. This would tell the the remote context to lookup the registry "cdn" off the container registry, and then ask that registry for the package registry, and then finally for the component factory.

Changes:

Remove ComponentFactoryTypes. This was for backwards compatibility which floating factory functions. these are not used anymore, so this can be removed.

Remove ComponentRegistryTypes. This was used to allow map to be used as registries. This however created problems, in that maps don't expose our feature detection interfaces. However, all public api's the took ComponentRegistryTypes, now take NamedComponentRegistires which is a iteratable, and map compatible.

Pass the component's IComponentRegistry to the component runtime, and make it available off IComponentRuntime. This allows easy access to the component's registry for feature discovery.