Properly propagate disconnected events on container / connection clos… (#617)
Properly propagate disconnected events on container / connection closure when underlying socket is left open for future reuse

We were never correctly raising disconnected event on closure of OdspDocumentDeltaConnection (IDocumentDeltaConnection). But because underlying socket was closed, we were getting "disconnect" event pretty quickly (though still asynchronously, and thus with some chance of hitting same problem).
With socket ODSP reuse in 0.11, we do not close socket for 2 seconds as result of connection closure, in an attempt to reuse same socket on reconnect. As result, disconnected events do not flow through the system, and DeltaManager is left in inconsistent state (no connection, but connected state). Bohemia attempts to submit an op in this state and we hit undefined dereference.

The fix is to make sure DeltaManager always stays in consistent state and raises disconnect events on connection closure. It already does it on disconnectFromDeltaStream() path, so reusing it.
Also making sure that OdspDocumentDeltaConnection raises connection on closure, and Container validates state is consistent on closure.