Replace 'random-name' with 'chance'
The npm 'random-name' package has a couple of issues that make it
unsuitable for use:
-It has an open issue where 6 random names are logged to the console of
the app as soon as the app opens.
-It mistakenly uses a list of (F)emale names as (F)irst names and (M)ale
name as (M)iddle names
-There are PRs to fix both issue submitted three years ago, but no
active owners to approve them.

The net result is console spam in every app using local-test-server.
I'm cleaning up console spam in ScriptorPad, the Scriptor test app. This
is now the largest source of console spam.

Replacing this library with 'chance', a test library for generating
random things. It's bigger than 'random-name', but that shouldn't be an
issue for local-test-server. And it is correct, currently maintained,
and doesn't spam the console.