Stop parsing in loadTardis to avoid double handle-parsing (#709)
Handles in the messages returned from loadTardis are getting handle-parsed before they are returned. Then in sequence.ts, these messages get handle-parsed again as they are passed through processMergeTreeMsg. This change removes the parse in loadTardis, such that Tardis only deals with messages in their serialized format (leaving it up to the consumer of the SnapshotLoader to process the messages in the normal fashion).