TestHost: Add scope option to testHost (#716)
Right now there is no way to manipulate the scope from an app using the testhost. This pr adds scope as a parameter of the constructor of testHost and passes it to the testDataStore via the open method.

This PR will enable us to pass in test app services for verification purposes.