Refactor flatten+buildHierarchy to buildSnapshotTree
The flatten function that convert an array of ITreeEntry to git.ITreeEntry.
We always use the output of that to buildHierarchy to convert it to
ISnapshotTree.  So refactor it so that we have a function to convert
directly to ITreeEntry to ISnapshotTree and avoid exposing git concepts
ot the runtime.