ESLint: Fully convert several @fluid-example packages to ESLint (#743)
This change fully converts the packages listed below to use ESLint rather than TSLint. After this change, these packages have no devDependency on TSLint.

To facilitate this, a new config was added to the shared ESLint config created in #430 called `@microsoft/eslint-config-fluid/no-tslint`. This config disables the [@typescript-eslint/eslint-plugin-tslint](https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin-tslint) plugin, which is what runs an instance of TSLint within ESLint so you get "the best of both worlds," so to speak. However, this has a perf cost, and TSLint itself will be full deprecated soon, so our goal is to _fully_ move to ESLint.

For each of the packages below, the following was done:

- Updated all code to address ESLint rules or ignore them
- Removed the tslint.json file from the project
- Updated eslintrc to use the no-tslint config
- (if needed) added an .editorconfig file to help VSCode format indentation properly
- Added eslint plugin dependencies to package.json and remove tslint dependency
- Updated tslint task
- Removed all // tslint comments
- Alphabetized rules in eslintrc

## Packages

Below is the list of packages that are changed in this commit:

- @fluid-example/draft-js
- @fluid-example/github-comment
- @fluid-example/markflow
- @fluid-example/math 
- @fluid-example/monaco
- @fluid-example/shared-map-visualizer
- @fluid-example/shared-text
- @fluid-example/table-view
- @fluid-example/tiny-web-host
- @fluid-example/todo
- @fluid-example/type-race
- @fluid-example/webflow
- @fluid-example/pollster