Convert connecting Deferred to a promise (#861)
This change consolidates the rejection logic into the connect method via the "closed" event. This then allows the Deferred to be made into a Promise, and resolve the floating promise exception. This change also ensures we fire the event before clearing the listeners, and cleans up a couple other eslint exceptions.