Add exception cases for string insertText ranges (#869)
* Add exception cases for string insertText ranges

Sequence's replaceRange doesn't handle cases where the start range is equal to or greater than the end range. Update the behavior not to run removal when the ranges are the same, and to always insert at the greater end of the range (which handles the case where the start > end of the range). Additionally for start > end, the removal step does not happen.