Update canvas control, remove FlexView and InkCanvas from client-ui (#836)
This change adds an InkCanvas control into the Ink package, derived from client-ui's InkCanvas. This control differs from the client-ui InkCanvas in that it removes client-ui dependencies and adds multitouch support, but otherwise is functionally similar to it.

This change also removes members from client-ui that have become dead code as a result, and removes the unused InkCanvas path from FlowContainer.